Refactor out client sentry code
Move it into `@/client/sentry`.
